feat: PostGIS 사용 기존 계산 로직 성능 최적화 진행

+++ src/main/java/com/project/safetyFence/geofence/GeofenceController.java
@@ -35,8 +35,8 @@ public class GeofenceController {
             Long id = geofence.getId();
             String name = geofence.getName();
             String address = geofence.getAddress();
-            BigDecimal latitude = geofence.getLatitude();
-            BigDecimal longitude = geofence.getLongitude();
+            BigDecimal latitude = BigDecimal.valueOf(geofence.getLatitude());
+            BigDecimal longitude = BigDecimal.valueOf(geofence.getLongitude());
             int type = geofence.getType();
             LocalDateTime startTime = geofence.getStartTime();
             LocalDateTime endTime = geofence.getEndTime();
